+++ psutil/_psutil_bsd.c
@@ -96,6 +96,9 @@
     #include <utmpx.h>
     #include <sys/vnode.h>  // for VREG
     #include <sys/sched.h>  // for CPUSTATES & CP_*
+    #ifndef DTYPE_VNODE
+    #define DTYPE_VNODE 1
+    #endif
 #endif
 
 
@@ -706,8 +709,10 @@ psutil_disk_partitions(PyObject *self, PyObject *args) {
             strlcat(opts, ",relatime", sizeof(opts));
         if (flags & MNT_IGNORE)
             strlcat(opts, ",ignore", sizeof(opts));
+#if defined(MNT_DISCARD)
         if (flags & MNT_DISCARD)
             strlcat(opts, ",discard", sizeof(opts));
+#endif
         if (flags & MNT_EXTATTR)
             strlcat(opts, ",extattr", sizeof(opts));
         if (flags & MNT_LOG)