+++ HISTORY.rst
@@ -13,10 +13,10 @@ Bug tracker at https://github.com/giampaolo/psutil/issues
   (patch by Eric Rahm)
 - #755: Process.memory_percent() "memtype" parameter.
 - #758: tests now live in psutil namespace.
-- #760: expose OS constants (LINUX, OSX, etc.)
+- #760: expose OS constants (psutil.LINUX, psutil.OSX, etc.)
 - #756: [Linux] disk_io_counters() return 2 new fields: read_merged_count and
   write_merged_count.
-- #762: new sripts/procsmem.py script.
+- #762: new scripts/procsmem.py script.
 
 **Bug fixes**