+++ HISTORY.rst
@@ -1,6 +1,6 @@
 Bug tracker at https://github.com/giampaolo/psutil/issues
 
-4.0.0 - XXXX-XX-XX
+4.0.0 - 2016-02-17
 ==================
 
 **Enhancements**